server/requirements.txt
RPi.GPIO==0.6.5
Flask==1.0.2
Flask-Cors==3.0.8
python-dotenv==0.10.3
